Moved version check
