+++ TCPChat-Server/MessageHandler.cs
@@ -114,7 +114,7 @@ namespace TCPChat_Server
             });
 
           // Check if server and client versions are the same before continuing.
-            if (!VersionCheck(instance, list[0].ClientVersion))
+            if (!VersionHandler.VersionCheck(instance, list[0].ClientVersion))
             {
                 // Remove the item just added to active clients.
                 // The reason it is added before is to have a list to index when sending server message to.
@@ -156,42 +156,6 @@ namespace TCPChat_Server
             }
 
             StreamHandler.WriteToStream(instance.stream, data);
-        }
-        public static bool VersionCheck(ClientInstance instance, string clientVersion)
-        {
-            // Check if versions do not match, if not close connection.
-            if (clientVersion != Assembly.GetExecutingAssembly().GetName().Version.ToString())
-            {
-                string message = String.Format("Your version of: {0} does not match the server version of: {1}",
-                    clientVersion,
-                    Assembly.GetExecutingAssembly().GetName().Version.ToString());
-
-
-                List<ServerMessageFormat> serverMessage = new List<ServerMessageFormat>();
-
-                serverMessage.Add(new ServerMessageFormat
-                {
-                    MessageType = MessageTypes.SERVER,
-                    Message = message,
-                    Color = ConsoleColor.Yellow,
-                    RSAExponent = Encryption.RSAExponent,
-                    RSAModulus = Encryption.RSAModulus,
-                });
-
-                string json = Serialization.Serialize(serverMessage);
-
-                byte[] data = Serialization.AddEndCharToMessage(json);
-
-                int index = MessageHandler.FindClientKeysIndex(instance.client);
-
-                byte[] encrypted = EncryptMessage(data, ServerHandler.activeClients[index].RSAModulus, ServerHandler.activeClients[index].RSAExponent);
-
-                StreamHandler.WriteToStream(instance.stream, encrypted);
-
-                instance.client.Close();
-                return false;
-            }
-            return true;
         }
         public static void ServerMessage(ConsoleColor color, string message)
         {
